Supprimer le quiz cadre
