+++ lib/screens/quiz_menu_screen.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import '../widgets/adaptive_appbar_title.dart';
 import '../widgets/ad_banner.dart';
 import '../widgets/modern_gradient_button.dart';
-import 'quiz_cadre_screen.dart';
 import 'quiz_pp_screen.dart';
 import 'quiz_stats_screen.dart';
 import 'recherche/recherche_infraction_list_screen.dart';
@@ -32,57 +31,42 @@ class QuizMenuScreen extends StatelessWidget {
                   child: Column(
                     mainAxisSize: MainAxisSize.min,
                     children: [
-                        FractionallySizedBox(
-                          widthFactor: 0.85,
-                          child: ModernGradientButton(
-                            icon: Icons.gavel,
-                            label: "Quiz cadres d'enquête",
-                            onPressed: () {
-                              Navigator.of(context).push(
-                                MaterialPageRoute(
-                                  builder: (_) => const QuizCadreScreen(),
-                                ),
-                              );
-                            },
-                          ),
+                      FractionallySizedBox(
+                        widthFactor: 0.85,
+                        child: ModernGradientButton(
+                          icon: Icons.school,
+                          label: 'Quiz PP',
+                          onPressed: () {
+                            Navigator.of(context).push(
+                              MaterialPageRoute(
+                                builder: (_) => const QuizPPScreen(),
+                              ),
+                            );
+                          },
                         ),
-                        const SizedBox(height: 16),
-                        FractionallySizedBox(
-                          widthFactor: 0.85,
-                          child: ModernGradientButton(
-                            icon: Icons.school,
-                            label: 'Quiz PP',
-                            onPressed: () {
-                              Navigator.of(context).push(
-                                MaterialPageRoute(
-                                  builder: (_) => const QuizPPScreen(),
-                                ),
-                              );
-                            },
-                          ),
+                      ),
+                      const SizedBox(height: 16),
+                      FractionallySizedBox(
+                        widthFactor: 0.85,
+                        child: ModernGradientButton(
+                          icon: Icons.help_outline,
+                          label: 'Quiz infractions',
+                          onPressed: () {
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(content: Text('À venir')),
+                            );
+                          },
                         ),
-                        const SizedBox(height: 16),
-                        FractionallySizedBox(
-                          widthFactor: 0.85,
-                          child: ModernGradientButton(
-                            icon: Icons.help_outline,
-                            label: 'Quiz infractions',
-                            onPressed: () {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                const SnackBar(content: Text('À venir')),
-                              );
-                            },
-                          ),
-                        ),
-                        const SizedBox(height: 16),
-                        FractionallySizedBox(
-                          widthFactor: 0.85,
-                          child: ModernGradientButton(
-                            icon: Icons.search,
-                            label: 'Recherche d’infractions',
-                            onPressed: () {
-                              Navigator.of(context).push(
-                                MaterialPageRoute(
+                      ),
+                      const SizedBox(height: 16),
+                      FractionallySizedBox(
+                        widthFactor: 0.85,
+                        child: ModernGradientButton(
+                          icon: Icons.search,
+                          label: 'Recherche d’infractions',
+                          onPressed: () {
+                            Navigator.of(context).push(
+                              MaterialPageRoute(
                                 builder: (_) =>
                                     const RechercheInfractionListScreen(),
                               ),
